gestione delle aggregazioni con interfaccia erlang, gestione del key value da nodo, ridefinizione dei package
